+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KREC7BWZGR74BTK8ECQYB9XZ"
+        "01KRED5HWW4NP988D0SZK89PA0"
     )
     r"""The unique identifier of the data source"""